+++ api/services/scoreService.ts
@@ -1,8 +1,5 @@
 import prisma from "../prismaInstance.js";
-import {
-  SaveScorePayload,
-  ScoresQueryOptions,
-} from "../types.js";
+import { SaveScorePayload, ScoresQueryOptions } from "../types.js";
 import { getR2Url } from "./r2.js";
 import {
   fetchBestScoreByPlayerId,
@@ -32,13 +29,13 @@ export const getHighScores = async () => {
     .map((s) => s.players?.profile_pic_r2_key)
     .filter((key) => typeof key === "string");
   const profilePicUrlMap = await createProfilePicUrlMap(r2Keys);
-  return res.map(({ players, points, level, outcomes, duration }) => ({
+  return res.map(({ players, points, level, outcome, duration }) => ({
     playerIdentifier: `${players?.external_id}-${players?.sso_provider}`,
     playerName: players?.name,
     playerProfilePicUrl: profilePicUrlMap[players?.profile_pic_r2_key ?? ""],
     points,
     level,
-    outcome: outcomes.desc,
+    outcome,
     duration,
   }));
 };
@@ -46,8 +43,8 @@ export const getHighScores = async () => {
 export const getPersonalBest = async (playerId: number) => {
   const res = await fetchBestScoreByPlayerId(prisma, playerId);
   if (res) {
-    const { points, level, duration, outcomes } = res;
-    return { points, level, duration, outcome: outcomes.desc };
+    const { points, level, duration, outcome } = res;
+    return { points, level, duration, outcome };
   } else {
     return null;
   }
@@ -69,12 +66,12 @@ export const getScores = async (options: ScoresQueryOptions) => {
     .map((s) => s.players?.profile_pic_r2_key)
     .filter((key) => typeof key === "string");
   const profilePicUrlMap = await createProfilePicUrlMap(r2Keys);
-  return res.map(({ players, points, level, outcomes, duration }) => ({
+  return res.map(({ players, points, level, outcome, duration }) => ({
     playerName: players?.name,
     playerProfilePicUrl: profilePicUrlMap[players?.profile_pic_r2_key ?? ""],
     points,
     level,
-    outcome: outcomes.desc,
+    outcome,
     duration,
   }));
 };